store signature of matching call with function reference

+++ lib/str_c.c
@@ -654,7 +654,7 @@ static void vsprt(char** s, unsigned* cap, unsigned* pos, const char* pre, const
 			double val;
 			if (ft->ID != VOLVOX_DoubleTyID) {
 				// C does not support variadic floats - so use some workaround
-				uint32_t u = va_arg(ap, uint32_t);
+				double u = va_arg(ap, double);
 				val = (double)*(float*)&u;
 			} else {
 				val = va_arg(ap, double);
